+++ src/components/AddShippingAddress.jsx
@@ -1,0 +1,144 @@
+import { useState } from "react";
+import { State, City } from "country-state-city";
+import useAppState from "../hooks/appState";
+import { useDispatch } from "react-redux";
+import { useForm } from "react-hook-form";
+import useApiMutation from "../api/hooks/useApiMutation";
+import { setKuduUser } from "../reducers/userSlice";
+
+const AddShippingAddress = ({ isOpen, closeModal, countries }) => {
+    const { user } = useAppState();
+    const dispatch = useDispatch();
+
+    const { mutate } = useApiMutation();
+
+
+    const { register, handleSubmit, formState: { errors }, reset, setValue } = useForm();
+
+
+    const [selectedCountry, setSelectedCountry] = useState(null);
+    const [selectedState, setSelectedState] = useState(null);
+    const [selectedCity, setSelectedCity] = useState(null);
+    const [states, setStates] = useState([]);
+    const [cities, setCities] = useState([]);
+
+    const handleCountryChange = (isoCode) => {
+        if (!isoCode) return;
+        const country = countries.find((c) => c.isoCode === isoCode);
+        setSelectedCountry(country);
+        setSelectedState(null);
+        setSelectedCity(null);
+        setStates(State.getStatesOfCountry(isoCode));
+        setCities([]);
+    };
+
+    const handleStateChange = (isoCode) => {
+        if (!isoCode || !selectedCountry) return;
+        
+        const state = states.find((s) => s.isoCode === isoCode);
+    
+        setSelectedState(state);
+        setSelectedCity(null);
+    
+        // Fetch cities using isoCode directly
+        const fetchedCities = City.getCitiesOfState(selectedCountry?.isoCode, isoCode);
+    
+        setCities(fetchedCities);
+    };
+    
+    const handleCityChange = (cityName) => {
+        if (!cityName || !selectedState) return;
+        setSelectedCity(cities.find((c) => c.name === cityName));
+    };
+
+    const onSubmit = (formData) => {
+        const payload = { firstName: user.firstName, lastName: user.lastName, dateOfBirth: null, gender: null, location: { ...formData } };
+
+        mutate({
+            url: "/user/profile/update",
+            method: "PUT",
+            data: payload,
+            headers: true,
+            onSuccess: (response) => {
+                dispatch(setKuduUser(response.data.data));
+                closeModal();
+            },
+            onError: (error) => {
+            },
+        });
+
+    };
+
+
+    return isOpen ? (
+        <form className="grid grid-cols-2 gap-1" onSubmit={handleSubmit(onSubmit)}>
+            {/* Country Selection */}
+            <div className="col-span-2">
+                <label className="block text-sm font-medium mt-4">Country</label>
+                <select
+                    className="w-full px-4 py-4 bg-gray-100 border border-gray-100 rounded-lg focus:outline-none text-sm mb-3"
+                    {...register("country", { required: "Country is required" })}
+                    onChange={(e) => handleCountryChange(e.target.value)}
+                >
+                    <option value="" disabled selected>Tap to Select</option>
+                    {countries.map((country) => (
+                        <option key={country.isoCode} value={country.isoCode}>
+                            {country.name}
+                        </option>
+                    ))}
+                </select>
+                {errors.country && <p className="text-red-500 text-sm">{errors.country.message}</p>}
+            </div>
+
+            {/* State Selection */}
+            <div className="col-span-2">
+                <label className="block text-sm font-medium mt-4">State</label>
+                <select
+                    className="w-full px-4 py-4 bg-gray-100 border border-gray-100 rounded-lg focus:outline-none text-sm mb-3"
+                    {...register("state", { required: "State is required" })}
+                    onChange={(e) => handleStateChange(e.target.value)}
+                    disabled={!selectedCountry}
+                >
+                    <option value="" disabled selected>Tap to Select</option>
+                    {states.map((state) => (
+                        <option key={state.isoCode} value={state.isoCode}>
+                            {state.name}
+                        </option>
+                    ))}
+                </select>
+                {errors.state && <p className="text-red-500 text-sm">{errors.state.message}</p>}
+            </div>
+
+            {/* City Selection */}
+            <div className="col-span-2">
+                <label className="block text-sm font-medium mt-4">City</label>
+                <select
+                    className="w-full px-4 py-4 bg-gray-100 border border-gray-100 rounded-lg focus:outline-none text-sm mb-3"
+                    onChange={(e) => handleCityChange(e.target.value)}
+                    {...register("city", { required: "City is required" })}
+                    disabled={!selectedState}
+                >
+                    <option value="" disabled selected>Tap to Select</option>
+                    {cities.map((city) => (
+                        <option key={city.name} value={city.name}>
+                            {city.name}
+                        </option>
+                    ))}
+                </select>
+                {errors.city && <p className="text-red-500 text-sm">{errors.city.message}</p>}
+            </div>
+
+            {/* Submit Button */}
+            <div className="col-span-2 mb-4 flex justify-start">
+                <button
+                    type="submit"
+                    className="bg-kuduOrange text-white text-sm font-medium py-4 px-4 rounded-md hover:bg-orange-600"
+                >
+                    Create Shipping Address
+                </button>
+            </div>
+        </form>
+    ) : null;
+};
+
+export default AddShippingAddress;
