+++ src/modules/User/modules/profile.jsx
@@ -1,18 +1,13 @@
 import React, { useEffect, useState } from "react";
 import useAppState from "../../../hooks/appState";
 import { useModal } from "../../../hooks/modal";
-import { City, Country, State } from "country-state-city";
-import { useForm } from "react-hook-form";
-import useApiMutation from "../../../api/hooks/useApiMutation";
-import { useDispatch } from "react-redux";
-import { setKuduUser } from "../../../reducers/userSlice";
+import { Country } from "country-state-city";
 import { useGetSubcriptionsPlanQuery } from "../../../reducers/storeSlice";
+import AddShippingAddress from "../../../components/AddShippingAddress";
 
 const AccountProfile = () => {
     const { user } = useAppState();
-    const dispatch = useDispatch();
     const { openModal, closeModal } = useModal();
-    const { register, handleSubmit, formState: { errors }, reset, setValue } = useForm();
 
     const [subscriptionPlan, setSubscriptionPlan] = useState([]);
 
@@ -29,147 +24,15 @@ const AccountProfile = () => {
 
     const countries = Country.getAllCountries();
 
-    const { mutate } = useApiMutation();
-
+    
     const handleModal = () => {
-        reset();
-        let selectedCountry = null;
-        let selectedState = null;
-        let selectedCity = null;
-        let states = [];
-        let cities = [];
-
-        const handleCountryChange = (isoCode) => {
-            if (!isoCode) return;
-            selectedCountry = countries.find((c) => c.isoCode === isoCode);
-            selectedState = null;
-            selectedCity = null;
-            states = State.getStatesOfCountry(selectedCountry.isoCode);
-            cities = [];
-
-            // Store the country name
-            setValue("country", selectedCountry.name);
-            renderModal();
-        };
-
-        const handleStateChange = (isoCode) => {
-            if (!isoCode || !selectedCountry) return;
-            selectedState = states.find((s) => s.isoCode === isoCode);
-            selectedCity = null;
-            cities = City.getCitiesOfState(selectedCountry.isoCode, selectedState.isoCode);
-
-            // Store the state name
-            setValue("state", selectedState.name);
-            renderModal();
-        };
-
-        const handleCityChange = (cityName) => {
-            if (!cityName || !selectedState) return;
-            selectedCity = cities.find((c) => c.name === cityName);
-
-            // Store the city name
-            setValue("city", selectedCity.name);
-            renderModal();
-        };
-
-
-
-        const onSubmit = (formData) => {
-            const payload = { firstName: user.firstName, lastName: user.lastName, dateOfBirth: null, gender: null, location: { ...formData } };
-
-            mutate({
-                url: "/user/profile/update",
-                method: "PUT",
-                data: payload,
-                headers: true,
-                onSuccess: (response) => {
-                    dispatch(setKuduUser(response.data.data));
-                    closeModal();
-                },
-                onError: (error) => {
-                },
-            });
-
-        };
-
-
-
-        const renderModal = () => {
-            openModal({
-                size: "sm",
-                content: (
-                    <form className="grid grid-cols-2 gap-1" onSubmit={handleSubmit(onSubmit)}>
-                        {/* Country Selection */}
-                        <div className="col-span-2">
-                            <label className="block text-sm font-medium mt-4">Country</label>
-                            <select
-                                className="w-full px-4 py-4 bg-gray-100 border border-gray-100 rounded-lg focus:outline-none text-sm mb-3"
-                                onChange={(e) => handleCountryChange(e.target.value)}
-                            >
-                                <option value="" disabled selected>Tap to Select</option>
-                                {countries.map((country) => (
-                                    <option key={country.isoCode} value={country.isoCode}>
-                                        {country.name}
-                                    </option>
-                                ))}
-                            </select>
-                            {errors.country && <p className="text-red-500 text-sm">{errors.country.message}</p>}
-                        </div>
-
-                        {/* State Selection */}
-                        <div className="col-span-2">
-                            <label className="block text-sm font-medium mt-4">State</label>
-                            <select
-                                className="w-full px-4 py-4 bg-gray-100 border border-gray-100 rounded-lg focus:outline-none text-sm mb-3"
-                                onChange={(e) => handleStateChange(e.target.value)}
-                                disabled={!selectedCountry}
-                            >
-                                <option value="" disabled selected>Tap to Select</option>
-                                {states.map((state) => (
-                                    <option key={state.isoCode} value={state.isoCode}>
-                                        {state.name}
-                                    </option>
-                                ))}
-                            </select>
-                            {errors.state && <p className="text-red-500 text-sm">{errors.state.message}</p>}
-                        </div>
-
-                        {/* City Selection */}
-                        <div className="col-span-2">
-                            <label className="block text-sm font-medium mt-4">City</label>
-                            <select
-                                className="w-full px-4 py-4 bg-gray-100 border border-gray-100 rounded-lg focus:outline-none text-sm mb-3"
-                                onChange={(e) => handleCityChange(e.target.value)}
-                                disabled={!selectedState}
-                            >
-                                <option value="" disabled selected>Tap to Select</option>
-                                {cities.map((city) => (
-                                    <option key={city.name} value={city.name}>
-                                        {city.name}
-                                    </option>
-                                ))}
-                            </select>
-                            {errors.city && <p className="text-red-500 text-sm">{errors.city.message}</p>}
-                        </div>
-
-                        {/* Submit Button */}
-                        <div className="col-span-2 flex justify-start">
-                            <button
-                                type="submit"
-                                className="bg-kuduOrange text-white text-sm font-medium py-4 px-4 rounded-md hover:bg-orange-600"
-                            >
-                                Create Shipping Address
-                            </button>
-                        </div>
-                    </form>
-                ),
-            });
-        };
-
-        renderModal();
-    };
-
-
+        openModal({
+            size: "sm",
+            content: (
+                <AddShippingAddress isOpen={true} countries={countries} closeModal={closeModal} />
+            )
+        })
+    }
